Move start activity helpers to activity module
