+++ sample/src/main/java/com/louiscad/splittiessample/main/MainActivity.kt
@@ -26,9 +26,9 @@ import com.louiscad.splittiessample.about.AboutActivity
 import com.louiscad.splittiessample.demo.DemoActivity
 import com.louiscad.splittiessample.extensions.menu.addItem
 import com.louiscad.splittiessample.extensions.menu.neverShowAsAction
-import com.louiscad.splittiessample.extensions.start
 import com.louiscad.splittiessample.extensions.toggleNightMode
 import com.louiscad.splittiessample.prefs.GamePreferences
+import splitties.activity.start
 import splitties.preferences.edit
 import splitties.snackbar.longSnack
 import splitties.systemservices.uiModeManager
